Barebones implementation of BoardPositionDiscarder.

+++ DiscardStrategy/ExpectedScores.h
@@ -19,7 +19,7 @@ public:
       int16_t opponent_play_mpoints = 0;
       int16_t opponent_hand_mpoints = 0;
 
-      int16_t observer_net_mpoints() const noexcept;
+      int observer_net_mpoints() const noexcept;
    };
    using Scores = std::array<Score, num_discard_actions>;
